Unify ID tracking. Apply flywheel design pattern to goods
r5040
2019-09-17

Note: this build aims to prep for 3.1.0 release.

## NEW
A. Unify the unit unique identifier and the resource ID
   (also call Good ID) :
1. Set the unit unique identifier in Unit as :
    - MARS_SURFACE_ID = 0
	- FIRST_SETTLEMENT_ID = 1
	- FIRST_BUILDING_ID = 20
    - FIRST_VEHICLE_ID = 2050
    - FIRST_PERSON_ID = 2500
    - FIRST_ROBOT_ID = 3550
    - FIRST_EQUIPMENT_ID = 4050
2. Set the resource ID (or Good ID) in ResourceUtil as :
    - FIRST_AMOUNT_RESOURCE_ID = 200
    - FIRST_ITEM_RESOURCE_ID = 1000
    - FIRST_VEHICLE_RESOURCE_ID = 2000
    - FIRST_EQUIPMENT_RESOURCE_ID = 4000
3. Create goodsMap in GoodsUtil for faster access.
4. Add 4 factors for modulating the VP of each type
   of vehicles :
   - TRANSPORT_VEHICLE_FACTOR = 1000D
   - CARGO_VEHICLE_FACTOR = 800D
   - EXPLORER_VEHICLE_FACTOR = 600D
   - LUV_VEHICLE_FACTOR = 200;

## CHANGE
1. Add a boolean to help the case when the build tag
   in version.txt is invalid.
2. Add 'h' and '/h' as keywords for LogMenu.
3. Add checking for marsSurface in getSettlement() in
   Equipment and Vehicle,
4. Modify getTopContainerUnit() in Unit.
5. Apply flywheel design pattern to streamline loading
   of goods:
   - Add goodsMap for getting goods
   - Replace calling createResourceGood() with
     getResourceGood()
   - Replace calling createEquipmentGood() with
     getEquipmentGood()
   - Replace calling createVehicleGood() with
     getVehicleGood()
6. Clean up Godd and GoodsUtil.
7. Move getFarmingAreaNeededPerPerson() from Farming to
   CropConfig.
8. Move averageCropGrowingTime from Crop to CropConfig.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/SimulationConfig.java
@@ -288,6 +288,7 @@ public class SimulationConfig implements Serializable {
 		String buildText = "";
 		boolean versionFileExist = false;
 		boolean xmlDirDeleted = false;
+		boolean invalid = false;
 		
 		// if the "xml" directory exists, back up everything inside and clean the directory
 		if (xmlDirExist && xmlLocation.isDirectory()) {
@@ -302,7 +303,7 @@ public class SimulationConfig implements Serializable {
 				    if ((buildText = brTest.readLine()) != null) {
 					    if (buildText.equals(Simulation.BUILD)) {	
 					    	sameBuild = true;
-					    }		
+					    }
 				    }
 				} catch (FileNotFoundException e) {
 					e.printStackTrace();
@@ -329,13 +330,15 @@ public class SimulationConfig implements Serializable {
 			LogConsolidated.log(Level.CONFIG, 0, sourceName, 
 					"Your version.txt already has the same BUILD " + buildText
 					+ " as the core engine.");
-		else if (!isNotNumber(buildText) )
+		else if (!isNotNumber(buildText))
 	    	LogConsolidated.log(Level.CONFIG, 0, sourceName, 
 					"Your version.txt has BUILD " + buildText 
 					+ ". The core engine has BUILD " + Simulation.BUILD);
-		else
+		else {
 			LogConsolidated.log(Level.CONFIG, 0, sourceName, 
 				"Your version.txt is invalid.");
+			invalid = true;
+		}
 		
 //		if (xmlDirExist && (!versionFileExist || !sameBuild)) {
 //			LogConsolidated.log(Level.CONFIG, 0, sourceName, 
@@ -348,13 +351,13 @@ public class SimulationConfig implements Serializable {
 			
 				try {
 	
-					if (versionFileExist && !buildText.equals("")) {
+					if (versionFileExist && !buildText.equals("") && !invalid) {
 						
 						String s0 = backupDir + File.separator + buildText;		
-				        File dir = new File(s0);
+				        File dir = new File(s0.trim());
 				        if (!dir.exists()) {
 				        	// Case A1 : Copy it to /.mars-sim/backup/buildText/
-							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case A2 : " +
+							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case A1 : " +
 									"Backing up to " + s0);
 							// Make a copy everything in the /xml to the /{$version}
 							FileUtils.copyDirectoryToDirectory(xmlLocation, dir);   	
@@ -365,8 +368,8 @@ public class SimulationConfig implements Serializable {
 				        	// Get timestamp in UTC
 				            Instant timestamp = Instant.now();
 				            String s1 = s0 + File.separator + timestamp.toString().replace(":", "").replace("-", "");
-				            dir = new File(s1);
-							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case A1 : " +
+				            dir = new File(s1.trim());
+							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case A2 : " +
 									s0 + " folder already exists. Backing up to " + s1);
 							// Make a copy everything in the /xml to the /{$version}
 							FileUtils.copyDirectoryToDirectory(xmlLocation, dir);
@@ -377,7 +380,7 @@ public class SimulationConfig implements Serializable {
 						
 						if (!backupLocation.exists()) {
 							// Case B1 : Copy it to /.mars-sim/backup/
-							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case B2 : " +
+							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case B1 : " +
 									"Backing up to " + backupDir);
 							// Make a copy everything in the /xml to the /backup/xml
 							FileUtils.copyDirectoryToDirectory(xmlLocation, backupLocation);
@@ -387,7 +390,7 @@ public class SimulationConfig implements Serializable {
 				            Instant timestamp = Instant.now();
 				            String s2 = backupDir + File.separator + timestamp.toString().replace(":", "").replace("-", "");
 				            backupLocation = new File(s2);
-							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case B1 : " +
+							LogConsolidated.log(Level.CONFIG, 0, sourceName, "Case B2 : " +
 									backupDir + " already exists. Backing up to " + s2);	
 							// Make a copy everything in the /xml to the /backup/xml
 							FileUtils.copyDirectoryToDirectory(xmlLocation, backupLocation);
